+++ src/se/osbe/emoreactor/brain/emotions/ConfusedEmotion.java
@@ -3,7 +3,7 @@ package se.osbe.emoreactor.brain.emotions;
 import se.osbe.emoreactor.brain.reactor.ReactorException;
 
 public class ConfusedEmotion extends AbstractEmotion {
-	public ConfusedEmotion(Double amplitude, long initialTime, long duration) throws ReactorException {
+	public ConfusedEmotion(Float amplitude, long initialTime, long duration) throws ReactorException {
 		super(EmotionType.CONFUSED, amplitude, initialTime, duration);
 	}
 }
