+++ src/se/osbe/emoreactor/brain/emotions/HelplessEmotion.java
@@ -3,7 +3,7 @@ package se.osbe.emoreactor.brain.emotions;
 import se.osbe.emoreactor.brain.reactor.ReactorException;
 
 public class HelplessEmotion extends AbstractEmotion {
-	public HelplessEmotion(Double amplitude, long initialTime, long duration) throws ReactorException {
+	public HelplessEmotion(Float amplitude, long initialTime, long duration) throws ReactorException {
 		super(EmotionType.HELPLESS, amplitude, initialTime, duration);
 	}
 }
