+++ src/se/osbe/emoreactor/brain/emotions/HurtEmotion.java
@@ -3,7 +3,7 @@ package se.osbe.emoreactor.brain.emotions;
 import se.osbe.emoreactor.brain.reactor.ReactorException;
 
 public class HurtEmotion extends AbstractEmotion {
-	public HurtEmotion(Double amplitude, long initialTime, long duration) throws ReactorException {
+	public HurtEmotion(Float amplitude, long initialTime, long duration) throws ReactorException {
 		super(EmotionType.HURT, amplitude, initialTime, duration);
 	}
 }
