+++ src/se/osbe/emoreactor/brain/emotions/IndifferentEmotion.java
@@ -3,7 +3,7 @@ package se.osbe.emoreactor.brain.emotions;
 import se.osbe.emoreactor.brain.reactor.ReactorException;
 
 public class IndifferentEmotion extends AbstractEmotion {
-	public IndifferentEmotion(Double amplitude, long initialTime, long duration) throws ReactorException {
+	public IndifferentEmotion(Float amplitude, long initialTime, long duration) throws ReactorException {
 		super(EmotionType.INDIFFERENT, amplitude, initialTime, duration);
 	}
 }
