+++ src/se/osbe/emoreactor/brain/emotions/PositiveEmotion.java
@@ -4,7 +4,7 @@ import se.osbe.emoreactor.brain.reactor.ReactorException;
 
 public class PositiveEmotion extends AbstractEmotion {
 	
-	public PositiveEmotion(Double amplitude, long initialTime, long duration) throws ReactorException {
+	public PositiveEmotion(Float amplitude, long initialTime, long duration) throws ReactorException {
 		super(EmotionType.POSITIVE, amplitude, initialTime, duration);
 	}
 }
