+++ src/se/osbe/emoreactor/brain/emotions/RelaxedEmotion.java
@@ -4,7 +4,7 @@ import se.osbe.emoreactor.brain.reactor.ReactorException;
 
 public class RelaxedEmotion extends AbstractEmotion {
 	
-	public RelaxedEmotion(Double amplitude, long initialTime, long duration) throws ReactorException {
+	public RelaxedEmotion(Float amplitude, long initialTime, long duration) throws ReactorException {
 		super(EmotionType.RELAXED, amplitude, initialTime, duration);
 	}
 }
